refactor(main): linked issue fetch helper を分離
起動 hydrate と PR 切替で共有する linked issue 並列 fetch を app::diff_viewer::linked_issues へ移す。

Refs: #224
