+++ src/main.rs
@@ -23,9 +23,8 @@ mod session;
 mod terminal;
 mod ui_state;
 
-use app::diff_viewer::snapshot::{
-    apply_snapshot_to_ui, build_pr_list_model, LinkedIssueDisplay,
-};
+use app::diff_viewer::linked_issues::fetch_linked_issues_parallel;
+use app::diff_viewer::snapshot::{apply_snapshot_to_ui, build_pr_list_model};
 use app::diff_viewer::refresh::{
     append_history, refresh_current_anchor_label, refresh_draft_panel, refresh_history_panel,
     refresh_preview, refresh_toasts,
@@ -35,9 +34,7 @@ use app::diff_viewer::state::{set_app_state, with_app_state, DiffAppState, Toast
 use app::diff_viewer::toast::{schedule_toast_auto_dismiss, set_active_window, show_toast};
 use app::diff_viewer::util::resolve_line_number;
 
-use github::issue_context::{
-    extract_linked_issue_numbers, fetch_issue_context_async,
-};
+use github::issue_context::extract_linked_issue_numbers;
 use github::pull_request::{
     build_client, fetch_pr_snapshot, fetch_pull_requests, parse_pr_spec, PrListFilter,
     PullRequestSnapshot,
@@ -1009,83 +1006,4 @@ fn wire_terminal_resize(
         // 通らず process exit) でも最後の window size を残せる (#231 補強)。
         save_window_session(&ui);
     });
-}
-
-/// linked issue 番号一覧を受け取り、各 issue を tokio::spawn 系で並列 fetch
-/// する。`octocrab::Octocrab` は内部で reqwest クライアントを共有しているので
-/// 数件の concurrent 呼び出しは安全。
-async fn fetch_linked_issues_parallel(
-    client: &octocrab::Octocrab,
-    owner: &str,
-    repo: &str,
-    numbers: &[u64],
-) -> Vec<LinkedIssueDisplay> {
-    use futures::stream::{FuturesUnordered, StreamExt};
-
-    let started = Instant::now();
-    let requested = numbers.len();
-    let mut futs: FuturesUnordered<_> = numbers
-        .iter()
-        .copied()
-        .map(|n| {
-            let client = client.clone();
-            let owner = owner.to_string();
-            let repo = repo.to_string();
-            async move {
-                let res = fetch_issue_context_async(&client, &owner, &repo, n).await;
-                (n, res)
-            }
-        })
-        .collect();
-
-    let mut out: Vec<LinkedIssueDisplay> = Vec::new();
-    let mut error_summary: Option<String> = None;
-    let mut error_count: usize = 0;
-    while let Some((n, res)) = futs.next().await {
-        match res {
-            Ok(Some(r)) => out.push(LinkedIssueDisplay::Found(r)),
-            Ok(None) => {}
-            Err(e) => {
-                let message = e.to_string();
-                if error_summary.is_none() {
-                    error_summary = Some(format!("#{n}: {message}"));
-                }
-                error_count += 1;
-                out.push(LinkedIssueDisplay::Failed {
-                    number: n,
-                    message,
-                });
-            }
-        }
-    }
-    // 並列実行の完了順は不定なので number でソートして決定論的にする
-    out.sort_by_key(|d| match d {
-        LinkedIssueDisplay::Found(r) => r.number,
-        LinkedIssueDisplay::Failed { number, .. } => *number,
-    });
-
-    tracing::debug!(
-        requested,
-        returned = out.len(),
-        error_count,
-        elapsed_ms = started.elapsed().as_millis() as u64,
-        "linked issues fetched"
-    );
-
-    // 1 件以上失敗していたら要約 toast を 1 つだけ出す。
-    // (各 issue chip は別途 error 表示されるため、toast は重複させない)
-    if let Some(first) = error_summary {
-        let count_str = error_count.to_string();
-        let _ = slint::invoke_from_event_loop(move || {
-            show_toast(
-                ToastKind::Warn,
-                i18n::tr_args(
-                    "Failed to fetch {} linked issue(s)",
-                    &[count_str.as_str()],
-                ),
-                first,
-            );
-        });
-    }
-    out
 }
